Cleaned up the form
- Define one widget for all form fields
- Need to update field values to model next
