--- shop_issue_form_flutter/lib/constants/custom_number_field.dart
@@ -1,36 +1,0 @@
-import 'dart:html';
-
-import 'package:flutter_form_builder/flutter_form_builder.dart';
-import 'package:flutter/material.dart';
-import 'package:form_builder_validators/form_builder_validators.dart';
-
-class CustomNumberField extends StatelessWidget {
-  final FocusNode focusName;
-  final String name;
-  final String labelTextName;
-  const CustomNumberField(
-      {Key? key,
-      required this.focusName,
-      required this.name,
-      required this.labelTextName})
-      : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return FormBuilderTextField(
-      focusNode: focusName,
-      name: name,
-      decoration: InputDecoration(
-        labelStyle: TextStyle(fontWeight: FontWeight.w500),
-        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
-        labelText: labelTextName,
-      ),
-      valueTransformer: (text) => num.tryParse(text!),
-      validator: FormBuilderValidators.compose([
-        FormBuilderValidators.required(context),
-        FormBuilderValidators.numeric(context),
-      ]),
-      keyboardType: TextInputType.number,
-    );
-  }
-}
