+++ shop_issue_form_flutter/packages/form_builder_file_picker/lib/src/form_builder_file_picker.dart
@@ -197,7 +197,6 @@ class _FormBuilderFilePickerState
 
     if (resultList != null) {
       setState(() => _files = [..._files, ...resultList!.files]);
-      // TODO: Pick only remaining number
       field.didChange(_files);
     }
   }
